Working version of a replacement for candle_db
Working / tested only for core="run".  No obvious code path in original for experiment_insert.  Should be a drop-in replacement.  Supplies default values for all but run_id (which is the unique key in solr anyway).  Solr seems to ignore field values with None, [], or "" so 'meaningful' default values should be specified.  (Still very chatty, lots of print staments left in.)